Enhance DTOs and exception handling: - Add custom message constructor to InvalidIdException. - Change price type from Float to BigDecimal in GameDtoResponse. - Introduce GamePatchDtoRequest for patching game details. - Update PurchaseDtoResponse to include lists of GameDtoResponse and UserDtoResponse. - Refactor PurchaseModel to adjust relationships with GameModel and UserModel. - Change UserController and PurchaseController to use @RestController annotation. - Update UserPatchDtoRequest and UserDtoRequest for improved validation messages. - Add handlers for HttpMessageNotReadableException and IllegalArgumentException in GlobalExceptionHandler.

+++ src/main/java/com/ludus/dtos/responses/GameDtoResponse.java
@@ -1,5 +1,6 @@
 package com.ludus.dtos.responses;
 
+import java.math.BigDecimal;
 import com.fasterxml.jackson.annotation.JsonIgnoreProperties;
 
 @JsonIgnoreProperties(ignoreUnknown = true)
@@ -9,7 +10,7 @@ public record GameDtoResponse(
   String genre,
   int releaseYear,
   String platform,
-  Float price
+  BigDecimal price
 ) {
 
 }
